refactor(YB-428): api 모듈 common 디렉토리 구조 재조정
