+++ yellobook-api/src/main/java/com/yellobook/common/aop/annotation/TeamAdminOnly.java
@@ -1,14 +1,11 @@
-package com.yellobook.common.annotation;
-
-import io.swagger.v3.oas.annotations.Parameter;
+package com.yellobook.common.aop.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Parameter(hidden = true)
-@Target(ElementType.PARAMETER)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface TeamMember {
+public @interface TeamAdminOnly {
 }
