+++ yellobook-api/src/main/java/com/yellobook/common/resolver/TeamMemberArgumentResolver.java
@@ -1,6 +1,6 @@
 package com.yellobook.common.resolver;
 
-import com.yellobook.common.annotation.TeamMember;
+import com.yellobook.common.resolver.annotation.TeamMember;
 import com.yellobook.common.vo.TeamMemberVO;
 import com.yellobook.domains.auth.security.oauth2.dto.CustomOAuth2User;
 import com.yellobook.domains.auth.service.RedisTeamService;
