+++ yellobook-api/src/main/java/com/yellobook/domains/member/controller/MemberController.java
@@ -1,6 +1,7 @@
 package com.yellobook.domains.member.controller;
 
-import com.yellobook.domains.auth.security.oauth2.dto.CustomOAuth2User;
+import com.yellobook.common.resolver.annotation.TeamMember;
+import com.yellobook.common.vo.TeamMemberVO;
 import com.yellobook.domains.member.dto.response.ProfileResponse;
 import com.yellobook.domains.member.service.MemberCommandService;
 import com.yellobook.domains.member.service.MemberQueryService;
@@ -10,7 +11,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -27,12 +27,11 @@ public class MemberController {
     @GetMapping("/profile")
     @Operation(summary = "마이프로필 조회", description = "로그인한 멤버의 마이프로필 조회 API")
     public ResponseEntity<SuccessResponse<ProfileResponse>> getMemberProfile(
-            @AuthenticationPrincipal CustomOAuth2User user
+            @TeamMember TeamMemberVO teamMember
     ) {
-        ProfileResponse response = memberQueryService.getMemberProfile(user.getMemberId());
-        return ResponseFactory.success(response);
+        var result = memberQueryService.getMemberProfile(teamMember.getMemberId());
+        return ResponseFactory.success(result);
     }
-
 
 
 //    @PatchMapping("/deactivate")
