+++ yellobook-api/src/main/java/com/yellobook/domains/team/controller/TeamController.java
@@ -1,6 +1,6 @@
 package com.yellobook.domains.team.controller;
 
-import com.yellobook.common.annotation.TeamMember;
+import com.yellobook.common.resolver.annotation.TeamMember;
 import com.yellobook.common.validation.annotation.ExistTeam;
 import com.yellobook.common.vo.TeamMemberVO;
 import com.yellobook.domains.auth.security.oauth2.dto.CustomOAuth2User;
